[Aura][ExplorerApi][Dev][ThuNPA] validators sync error
(cherry picked from commit 04baa0ea0314d2e34f0593dde6c95b1590c0c42a)

+++ src/shared/entities/validator.entity.ts
@@ -39,7 +39,7 @@ export class Validator extends BaseEntityIncrementId {
   @Column({ type: 'text' })
   delegator_shares: string;
 
-  @Column({ type: 'float' })
+  @Column({ type: 'double' })
   power: number;
 
   @Column({ default: '' })
